+++ lib/rubocop/cop/style/rbs_inline/data_define_with_block.rb
@@ -30,10 +30,11 @@ module RuboCop
         class DataDefineWithBlock < Base
           prepend FileFilter
           include DataClassMatcher
+          include ASTUtils
 
           MSG = "Do not use `Data.define` with a block. RBS::Inline does not parse block contents, " \
-                "so methods defined in the block will not be recognized. " \
-                "Use a separate class definition instead."
+                "so methods defined in the block will not be recognized. Keep the `Data.define` call " \
+                "and move the methods into %<destination>s."
 
           # @rbs node: RuboCop::AST::SendNode
           def on_send(node) #: void
@@ -42,7 +43,18 @@ module RuboCop
             block_node = node.parent
             return unless block_node&.block_type?
 
-            add_offense(node)
+            add_offense(node, message: format(MSG, destination: destination(block_node)))
+          end
+
+          private
+
+          # The constant the definition is assigned to is the class to reopen.
+          # @rbs node: RuboCop::AST::Node
+          def destination(node) #: String
+            class_name = assigned_constant_name(node)
+            return "a `class` that reopens it" unless class_name
+
+            "`class #{class_name} ... end` written after it"
           end
         end
       end